refactor: assets 내부의 이미지들을 디렉토리 안에 구조화
