+++ src/components/Header/Header.tsx
@@ -1,5 +1,5 @@
 import styles from './Header.module.css';
-import logo from '@/assets/logo.svg';
+import logo from '@/assets/common/logo.svg';
 
 function Header() {
   return (
